+++ types/validators/customValidators/index.d.ts
@@ -34,6 +34,11 @@ declare const _default: {
             valid: boolean;
         };
     };
+    isPhoneNumberJp: () => {
+        validate: (input: FormValidation.core.ValidateInput<FormValidation.core.ValidateOptions, FormValidation.core.Localization>) => {
+            valid: boolean;
+        };
+    };
     isPort: () => {
         validate: (input: FormValidation.core.ValidateInput<FormValidation.core.ValidateOptions, FormValidation.core.Localization>) => {
             valid: boolean;